==> code/Notebooks/01_CleaningCarMax_Data.ipynb <==
# Cleaning Car Max Web Scraped Data 

#### Here we simply clean the scraped CarMax data, so we can join it to the kaggle data

import pandas as pd

# load the data
car_data = pd.read_csv('Raw_CarMax.csv')

# clean 'car price' column by removing dollar signs, commas, asterisks, and extra spaces, then convert to float
car_data['Car Price'] = car_data['Car Price'].replace({'\$': '', ',': '', '\*': '', ' ': ''}, regex=True)

# display the df
display(car_data)

# remove 'k mi' and convert to a number in thousands
car_data['Car Mileage'] = car_data['Car Mileage'].str.replace('K mi', '').astype(float) * 1000

# format the 'car mileage' column with commas 
car_data['Car Mileage'] = car_data['Car Mileage'].apply(lambda x: f"{x:,.0f}")

# show the updated df
display(car_data)

# remove commas from the 'car mileage' column and convert to integers
car_data['Car Mileage'] = car_data['Car Mileage'].str.replace(',', '').astype(int)

# drop the 'car trim' column
car_data.drop('Car Trim', axis=1, inplace=True)

# display the cleaned df
display(car_data)

# save the cleaned DF
car_data.to_csv('Cleaned_CarMax.csv', index=False)
print("Data saved to 'Cleaned_CarMax.csv'")
