==> code/Notebooks/02_DataJoin&KaggleClean.ipynb <==
# Sports Car Prices Project - Kaggle Cleaning & Data Combination

### Max Enabnit, Greg Fagan, Luke Rubin


This notebook cleans the Sport car price Kaggle dataset and combines it with the scraped clean CarMax dataset, imputing averages for NaN values. It then exports the cleaned data to be used in the next notebook.

Site used to scrape CarMax car data:
https://www.carmax.com/cars/sports-cars

Kaggle Raw Data Link: https://www.kaggle.com/datasets/rkiattisak/sports-car-prices-dataset/data

import pandas as pd

# Load raw kaggle csv as a pandas data frame
cars_df = pd.read_csv('Raw_Kaggle.csv')

# Remove unnecessary columns
cars_df = cars_df.drop(columns="Torque (lb-ft)")

display(cars_df)

# Change the column names
cars_df = cars_df.rename(columns={
    "Car Make": "car_make",
    "Car Model": "car_model",
    "Year": "year",
    "Engine Size (L)": "engine_size",
    "Horsepower": "horsepower",
    "0-60 MPH Time (seconds)": "acceleration_time",
    "Price (in USD)": "price"
})

display(cars_df)

# Remove instances of electric cars and cars that have NaN engine_size
electric = cars_df[cars_df['engine_size'].str.contains('electric', case = False, na = True)]
cars_df = cars_df.drop(electric.index)

# Remove instances of hybrid cars
hybrid = cars_df[cars_df['engine_size'].str.contains('hybrid', case = False, na = False)]
cars_df = cars_df.drop(hybrid.index)

# Remove instances of cars that have '-' and '0' engine_size
null = cars_df[cars_df['engine_size'] == '-']
cars_df = cars_df.drop(null.index)

zeroes = cars_df[cars_df['engine_size'] == '0']
cars_df = cars_df.drop(zeroes.index)

cars_df

# Check data types of columns
cars_df.dtypes

# Change columns to correct data types
cars_df['car_make'] = cars_df['car_make'].astype('category')
cars_df['car_model'] = cars_df['car_model'].astype('category')
cars_df['engine_size'] = cars_df['engine_size'].astype('float')
cars_df['acceleration_time'] = cars_df['acceleration_time'].astype('float')

# Remove commas in horsepower price to change data type
cars_df['horsepower'] = cars_df['horsepower'].str.replace(',','')
cars_df['price'] = cars_df['price'].str.replace(',','')

# Change columns to correct data types
cars_df['horsepower'] = cars_df['horsepower'].astype('int')
cars_df['price'] = cars_df['price'].astype('int')

cars_df.dtypes

# Export cleaned data to a csv
cars_df.to_csv('Cleaned_Kaggle.csv', index=False)

# Import scraped carmax data set as a pandas data frame
carmax = pd.read_csv('Cleaned_CarMax.csv')
carmax

# Rename columns
carmax = carmax.rename(columns={
    "Car Make": "car_make",
    "Car Model": "car_model",
    "Car Year": "year",
    "Car Mileage": "mileage",
    "Car Price": "price"
})
carmax.dtypes

# Change data types
carmax['car_make'] = carmax['car_make'].astype('category')
carmax['car_model'] = carmax['car_model'].astype('category')
carmax['mileage'] = carmax['mileage'].astype('int')

# Merge the dataframes to create the sportscars dataframe
sportscars = pd.concat([cars_df, carmax], ignore_index=True)
sportscars

# Select columns with NaN values to be filled by similar cars' averages
columns_to_fill = ['engine_size', 'horsepower', 'acceleration_time']

# Calculate the average for each car model
car_model_means = sportscars.groupby('car_model')[columns_to_fill].mean()

# Fill NaN values with the calculated means for each car model
for column in columns_to_fill:
    sportscars[column] = sportscars.apply(
        lambda row: car_model_means.loc[row['car_model'], column] if pd.isna(row[column]) else row[column], axis=1
    )

# Calculate the average for each car make
car_make_means = sportscars.groupby('car_make')[columns_to_fill].mean()

# Fill the rest of the NaN values with the calculated means for each car make
for column in columns_to_fill:
    sportscars[column] = sportscars.apply(
        lambda row: car_make_means.loc[row['car_make'], column] if pd.isna(row[column]) else row[column], axis=1
    )

# Group by 'Car Year' and calculate the mean 'Car Mileage' for each year
car_year_means = sportscars.groupby('year')['mileage'].mean()

# Replace NaN values in 'Car Mileage' with the average mileage of the corresponding car year
sportscars['mileage'] = sportscars.apply(
    lambda row: car_year_means[row['year']] if pd.isna(row['mileage']) else row['mileage'], axis=1
)
sportscars

# Drop remaining NaN columns
sportscars=sportscars.dropna()

# Round values
sportscars['horsepower'] = sportscars['horsepower'].round()
sportscars['acceleration_time'] = sportscars['acceleration_time'].round(2)
sportscars['mileage'] = sportscars['mileage'].round()
sportscars['engine_size'] = sportscars['engine_size'].round()
sportscars

# Export the final combined data frame as a csv
sportscars.to_csv('Combined.csv', index=False)
